add xlnet

+++ model.py
@@ -48,7 +48,7 @@ class AvgPoolClassifier(CLSClassifier):
             x (torch.FloatTensor): logits of shape (batch_size, NUM_CLASS)
         
         """
-        x_mask = sequence_mask(length, pad=0)  # (batch_size, max_length)
+        x_mask = sequence_mask(length, pad=0, dtype=torch.float)  # (batch_size, max_length)
         # TODO: clean this hack
         try: # bert, roberta
             x, _ = self.model(x, attention_mask=x_mask)                     # (batch_size, seq_length, hidden_size)
